requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-docker
pyspark==3.2.1
docker==5.0.0
